fix 3.8 pickle issue on Mac
